Automatic reconnect: honour the automatic reconnect configuration, even if the GUI interface is open
When a user actives "automatic reconnnect" and then tries this out, he will be staring at the GUI interface, waiting for a reconnect.
But wicd-daemon refuses to reconnect when the GUI interface is open, and fails to reports this to the GUI user.
Hence users assume that this feature does not work properly (see https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=582567).

This commit fixes that by doing a reconnect, no matter whether the GUI is open or not, thus taking the path of "least surprise".
The original rationale for not reconnecting when the GUI is open was never stated in the comments, nor in the commit message.